requirements.txt
Cython==0.29.32
debugpy==1.6.3
decorator==5.1.1
defusedxml==0.7.1
Deprecated==1.2.13
Django==3.2.13
django-adminlte-3==0.1.6
django-crispy-forms==1.14.0
django-dotenv==1.4.2
django-environ==0.9.0
django-filter==22.1
django-taggit==3.0.0
django-templated-mail==1.1.1
djangorestframework==3.13.1
djangorestframework-simplejwt==4.7.2
dnspython==2.2.1
docker==6.0.0
dominate==2.7.0
ecdsa==0.18.0
email-validator==1.3.0
entrypoints==0.4
environ==1.0
fake-useragent==0.1.11
fastapi==0.66.0
flake8==4.0.1
flake8-broken-line==0.4.0
flake8-plugin-utils==1.3.2
flake8-polyfill==1.0.2
flake8-return==1.1.3
Flask==2.1.3
Flask-Bootstrap==3.3.7.1
Flask-GraphQL==2.0.1
Flask-GraphQL-Auth==1.3.3
Flask-JWT-Extended==3.8.1
Flask-Login==0.6.2
Flask-Mail==0.9.1
Flask-Migrate==3.1.0
Flask-RESTful==0.3.9
Flask-SQLAlchemy==2.5.1
Flask-WTF==1.0.1
graphene==2.1.8
graphene-sqlalchemy==2.3.0
graphql-core==2.3.2
graphql-relay==2.0.1
graphql-server-core==1.2.0
greenlet==1.1.0
gunicorn==20.0.4
h11==0.12.0
httpcore==0.13.7
httpx==0.18.2
idna==3.3
importlib-resources==5.9.0
iniconfig==1.1.1
install==1.3.5
ipykernel==6.15.3
ipython==7.34.0
is-disposable-email==1.0.0
isort==5.10.1
itsdangerous==2.1.2
itypes==1.2.0
jedi==0.18.1
Jinja2==3.1.2
jupyter-core==4.11.1
keycloak==3.1.3
kombu==5.2.4
load-dotenv==0.1.0
lxml==4.9.1
Mako==1.1.4
MarkupSafe==2.1.1
matplotlib-inline==0.1.6
mccabe==0.6.1
mypy-extensions==0.4.3
nest-asyncio==1.5.5
numpy==1.21.6
oauthlib==3.2.0
outcome==1.2.0
packaging==21.2
pandas==1.3.5
parso==0.8.3
passlib==1.7.4
pathspec==0.8.1
pep8-naming==0.12.1
pickleshare==0.7.5
Pillow==9.2.0
pluggy==0.13.1
promise==2.3
prompt-toolkit==3.0.31
psutil==5.9.2
psycopg2==2.9.1
psycopg2-binary==2.8.6
py==1.10.0
py3dns==3.2.1
pyasn1==0.4.8
pycodestyle==2.8.0
pycparser==2.21
pydantic==1.8.2
pyflakes==2.4.0
Pygments==2.13.0
PyJWT==2.0.1
PyMySQL==1.0.2
pyparsing==2.4.7
PySocks==1.7.1
pytest==6.2.4
pytest-django==4.4.0
pytest-pythonpath==0.7.3
python-decouple==3.6
python-dotenv==0.18.0
python-editor==1.0.4
python-jose==3.3.0
python3-openid==3.2.0
pytz==2021.3
pywin32==304
PyYAML==5.4.1
pyzmq==24.0.0
redis==4.3.4
regex==2021.7.1
requests==2.28.1
requests-oauthlib==1.3.1
requests-toolbelt==0.9.1
rfc3986==1.5.0
rsa==4.9
Rx==1.6.1
selenium==4.4.3
selenium-stealth==1.0.6
singledispatch==3.6.2
six==1.16.0
sniffio==1.2.0
social-auth-app-django==4.0.0
social-auth-core==4.3.0
sortedcontainers==2.4.0
soupsieve==2.3.2.post1
SQLAlchemy==1.4.20
sqlparse==0.4.2
starlette==0.14.2
stripe==3.5.0
style==1.1.0
toml==0.10.2
tornado==6.2
traitlets==5.4.0
trio==0.21.0
trio-websocket==0.9.2
typed-ast==1.5.4
typing-extensions==3.10.0.0
undetected-chromedriver==3.1.5.post4
update==0.0.1
uritemplate==4.1.1
urllib3==1.26.11
uvicorn==0.14.0
validate-email==1.3
validators==0.20.0
vine==5.0.0
visitor==0.1.3
wcwidth==0.2.5
websocket-client==1.4.1
websockets==10.3
Werkzeug==2.2.2
windows-curses==2.3.0
wrapt==1.14.1
wsproto==1.2.0
WTForms==3.0.1
wtforms-validators==1.0.0
zipp==3.8.1
